Refactor update code into classes
